+++ app/home/historical/page.tsx
@@ -12,66 +12,28 @@ export default async function Page(props: {
   }>;
   }) {
     const searchParams = await props.searchParams;
-    // console.log("searchparams...",searchParams)
     const currentPage = Number(searchParams?.page) || 1;
-    // console.log("currentPage...",currentPage)
     const query = searchParams?.query || '';
     console.log("query...",query)
     const lighthouses = await fetchLighthouses(currentPage, query);
     const totalPages = await fetchLighthousePages(query);
-    // console.log("totalpages....",totalPages)
-    // console.log("fetch....",lighthouses)
+   
   return (
     <>
       <h1>Historical Weather</h1>
+      <Search placeholder="Search Lighthouse..." />
       {lighthouses.map(async(lighthouse) => {
         const { daily } = await fetchHistoricalWeather(lighthouse.latitude, lighthouse.longitude, 'Europe/Dublin' );
-        // console.log("Current....",daily);
-        // console.log(lighthouse.name);
-
+        
         return (
           <>
             <h1>{ lighthouse.name }</h1>
             <div>  
               <LineGraph daily = {daily}/>
             </div>
-            {/* <Header
-              currentTemp={current?.currentTemp}
-              highTemp={current?.highTemp}
-              lowTemp={current?.lowTemp}
-              highFeelsLike={current?.highFeelsLike}
-              lowFeelsLike={current?.lowFeelsLike}
-              windSpeed={current?.windSpeed}
-              precip={current?.precip}
-              iconCode={current?.iconCode}
-            /> */}
-            {/* <section className="grid grid-cols-[repeat(auto-fit,100px)] gap-2 justify-center p-4">
-              {daily.map((item, index) => (
-                <DayCard
-                  key={index}
-                  iconCode={item.iconCode}
-                  timestamp={item.timestamp}
-                  degree={item.maxTemp}
-                />
-              ))}
-            </section> */}
-             {/* <table className="w-full text-center border-spacing-0">
-                <tbody>
-                  {daily.map((item, index) => (
-                    <TableRow
-                      key={index}
-                      maxWind={item.maxWind}
-                      timestamp={item.timestamp}
-                      maxGust={item.maxGust}
-                    
-                    />
-                  ))}
-                </tbody>
-            </table> */}
           </>
         )
       })}
-      <Search placeholder="Search Lighthouse..." />
       <div className="mt-5 flex w-full justify-center">
         <Pagination totalPages={totalPages} />
       </div>